auth/requirements.txt
fastapi==0.68.0
sqlalchemy==1.4.22
psycopg2-binary==2.9.1
pydantic==1.8.2
uvicorn[standard]==0.14.0
bcrypt==3.2.0
pydantic[email]
email-validator===2.1.1
python-jose
passlib
python-multipart
databases
httpx==0.23.0
requests
